+++ src/main/java/com/mageddo/conf/parser/Entry.java
@@ -1,0 +1,17 @@
+package com.mageddo.conf.parser;
+
+import lombok.Builder;
+import lombok.NonNull;
+import lombok.Value;
+
+@Value
+@Builder
+public class Entry {
+
+  @NonNull
+  EntryType type;
+
+  @NonNull
+  String line;
+
+}