+++ src/main/java/com/mageddo/conf/parser/EntryType.java
@@ -1,0 +1,10 @@
+package com.mageddo.conf.parser;
+
+public interface EntryType {
+
+  String name();
+
+  static EntryType of(String name) {
+    return new EntryTypeDefault(name);
+  }
+}