+++ src/main/java/com/mageddo/conf/parser/Transformer.java
@@ -1,0 +1,12 @@
+package com.mageddo.conf.parser;
+
+import java.util.Set;
+
+public interface Transformer {
+
+  String handle(Entry entry);
+
+  default String after(boolean fileHasContent, Set<String> foundEntryTypes) {
+    return null;
+  }
+}