+++ src/main/java/com/mageddo/dnsproxyserver/config/entrypoint/ConfigEnv.java
@@ -26,7 +26,8 @@ public class ConfigEnv {
   public static final String MG_DOMAIN = "MG_DOMAIN";
   public static final String MG_DPS_NETWORK = "MG_DPS_NETWORK";
   public static final String MG_DPS_NETWORK_AUTO_CONNECT = "MG_DPS_NETWORK_AUTO_CONNECT";
-  public static final String DEFAULT_RESOLV_CONF_PATH = "/host/etc/resolv.conf,/etc/resolv.conf";
+  public static final String DEFAULT_RESOLV_CONF_PATH =
+    "/host/etc/systemd/resolved.conf,/host/etc/resolv.conf,/etc/systemd/resolved.conf,/etc/resolv.conf";
 
   private Path currentPath;
   private String resolvConfPath;