+++ src/main/java/com/mageddo/dnsproxyserver/dnsconfigurator/linux/LinuxDnsConfigurator.java
@@ -3,10 +3,10 @@ package com.mageddo.dnsproxyserver.dnsconfigurator.linux;
 import com.mageddo.commons.lang.Objects;
 import com.mageddo.dnsproxyserver.config.Configs;
 import com.mageddo.dnsproxyserver.dnsconfigurator.DnsConfigurator;
-import com.mageddo.dnsproxyserver.dnsconfigurator.linux.resolvconf.DnsServerCleanerHandler;
-import com.mageddo.dnsproxyserver.dnsconfigurator.linux.resolvconf.SetMachineDNSServerHandler;
-import com.mageddo.dnsproxyserver.resolvconf.ResolvConfParser;
+import com.mageddo.dnsproxyserver.dnsconfigurator.linux.ResolvFile.Type;
 import com.mageddo.dnsproxyserver.server.dns.IP;
+import com.mageddo.dnsproxyserver.systemd.ResolvedService;
+import com.mageddo.utils.Tests;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -17,15 +17,19 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
 
 import static com.mageddo.dnsproxyserver.utils.Splits.splitToPaths;
+import static org.apache.commons.lang3.ObjectUtils.firstNonNull;
 
 @Slf4j
 @Default
 @Singleton
 @RequiredArgsConstructor(onConstructor = @__({@Inject}))
 public class LinuxDnsConfigurator implements DnsConfigurator {
+
+  private final AtomicBoolean resolvedConfigured = new AtomicBoolean();
 
   private volatile AtomicReference<ResolvFile> confFile;
 
@@ -37,7 +41,15 @@ public class LinuxDnsConfigurator implements DnsConfigurator {
       return;
     }
 
-    ResolvConfParser.process(getConfFile(), new SetMachineDNSServerHandler(ip.raw()));
+    final var confFile = this.getConfFile();
+    if (confFile.isResolvconf()) {
+      ResolvconfConfigurator.process(confFile.getPath(), ip);
+    } else if (confFile.isResolved()) {
+      this.configureResolved(ip, confFile);
+    } else {
+      throw newUnsupportedConfType(confFile);
+    }
+    log.debug("status=configured, path={}", this.getConfFile());
   }
 
   @Override
@@ -46,12 +58,21 @@ public class LinuxDnsConfigurator implements DnsConfigurator {
     if (this.confFile.get() == null) {
       return;
     }
-    ResolvConfParser.process(getConfFile(), new DnsServerCleanerHandler());
-    log.debug("status=restoredResolvConf, path={}", this.getConfFile());
+
+    final var confFile = this.getConfFile();
+    if (confFile.isResolvconf()) {
+      ResolvconfConfigurator.restore(confFile.getPath());
+    } else if (confFile.isResolved()) {
+      ResolvedConfigurator.restore(confFile.getPath());
+      tryRestartResolved();
+    } else {
+      throw newUnsupportedConfType(confFile);
+    }
+    log.debug("status=restored, path={}", this.getConfFile());
   }
 
-  Path getConfFile() {
-    return this.confFile.get().getPath();
+  ResolvFile getConfFile() {
+    return this.confFile.get();
   }
 
   ResolvFile findBestConfFile() {
@@ -87,13 +108,40 @@ public class LinuxDnsConfigurator implements DnsConfigurator {
   }
 
   ResolvFile toResolvFile(Path path) {
-    return ResolvFile.of(path, LinuxResolverConfDetector.detect(path));
+    return ResolvFile.of(path, firstNonNull(LinuxResolverConfDetector.detect(path), Type.RESOLVCONF));
   }
 
   void init() {
     if (this.confFile == null) {
-      this.confFile = new AtomicReference<>(findBestConfFile());
+      this.confFile = new AtomicReference<>(this.findBestConfFile());
       log.info("status=using, configFile={}", this.getConfFile());
+    }
+  }
+
+  private RuntimeException newUnsupportedConfType(ResolvFile confFile) {
+    return new UnsupportedOperationException(String.format("conf file not supported: %s", confFile));
+  }
+
+  private void configureResolved(IP ip, ResolvFile confFile) {
+    if (this.resolvedConfigured.compareAndSet(false, true)) {
+      ResolvedConfigurator.configure(confFile.getPath(), ip);
+      tryRestartResolved();
+    }
+  }
+
+  static void tryRestartResolved() {
+    try {
+      if (Tests.inTest()) {
+        log.warn("status=wont-restart-service-while-testing");
+        return;
+      }
+      ResolvedService.restart();
+    } catch (Throwable e) {
+      log.warn(
+        "status=can't restart resolved service, please run: "
+          + "'service systemd-resolved restart' to apply DPS as default DNS.\n{}",
+        e.getMessage()
+      );
     }
   }