+++ src/main/java/com/mageddo/dnsproxyserver/quarkus/Quarkus.java
@@ -1,5 +1,6 @@
 package com.mageddo.dnsproxyserver.quarkus;
 
+import com.mageddo.utils.Tests;
 import io.quarkus.arc.Arc;
 import io.quarkus.runtime.configuration.ConfigUtils;
 
@@ -9,16 +10,17 @@ import java.util.stream.Collectors;
 public class Quarkus {
   public static boolean isTest() {
     return ConfigUtils
-        .getProfiles()
-        .contains("test")
-        ;
+      .getProfiles()
+      .contains("test")
+      || Tests.inTest()
+      ;
   }
 
   public static <T> Set<T> beansOf(Class<T> clazz) {
     return Arc.container()
-        .select(clazz)
-        .stream()
-        .collect(Collectors.toSet())
-        ;
+      .select(clazz)
+      .stream()
+      .collect(Collectors.toSet())
+      ;
   }
 }