+++ src/main/java/com/mageddo/utils/Tests.java
@@ -1,7 +1,7 @@
 package com.mageddo.utils;
 
 public class Tests {
-  public static boolean runningOnJunit() {
+  public static boolean inTest() {
     for (StackTraceElement element : Thread.currentThread().getStackTrace()) {
       if (element.getClassName().startsWith("org.junit.")) {
         return true;